feat(dbus-ibr-system): Publish battery info as string array
Change the DBus interface for battery information. Instead of publishing
device name and BT MAC address on separate paths for each device,
a single path `/Info/BattInfo` is created.

This path holds a flat list of strings with `[device, devicename, btmac, ...]`
for all configured batteries. This makes it easier for clients to get all
information at once and reduces the number of DBus paths.