feat(serialbat): Use BattInfo for ProductName
Update dbus-ibr-serialbat to use the /Info/BattInfo D-Bus path
from com.victronenergy.ibrsystem to set a more descriptive ProductName.

The ProductName now includes the battery's friendly name if available
from the BattInfo.

This commit also fixes a NameError that could occur if the battery's
information was not found in BattInfo.